Created add book and update book
Create a add new book button so you can manually enter a new book, improvement in the future for this will be to use some of of book API maybe waterstones? to go and fetch the book data through a search bar instead of having to upload it manually. and a update book page where you can update the details of the book