eventcard functionlity, calendar functionality, monthselector funationality. Event card overlay added
alot of progress made today, the font page of the app now works as you think it would with the month selector working and displaying the correct month, the calendar displaying the correct layout for the month e.g. correct number of days and days of the week. the enventcard overlay now works to click into the event to see the details. and can click off the overlay anywhere on the screen. Next steps will be funtionality of the reminder and notification buttons and making the ui for the files page. further further steps will be to work on the profile page.