deleted useless projects and added new project benefit checker
benefit checker is a full stack web application designed to access a persons eligibility for government benefits based on their income, family size and location. Ive made this project to practise java and spring boot as well as making a rest api. this project is my first real attempt at a backend project, so its pretty simple so far. the front end is just a simple react form.